+++ armonaut/cache/http.py
@@ -12,9 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import collections.abc
 import functools
 
 __all__ = ['add_vary']
+
+BUFFER_MAX = 1024 * 1024
 
 
 def add_vary_callback(*varies):
@@ -33,3 +36,40 @@ def add_vary(*varies):
             return view(context, request)
         return wrapped
     return wrapper
+
+
+def conditional_http_tween_factory(handler, registry):
+    def conditional_http_tween(request):
+        response = handler(request)
+
+        # If the Last-Modified header has been set enable
+        # conditional response handling.
+        if response.last_modified is not None:
+            response.conditional_response = True
+
+        # Only enable conditional response handling for ETag if
+        # we're given an ETag or we have a buffered response and
+        # can generate the ETag header ourselves.
+        if response.etag is not None:
+            response.conditional_response = True
+
+        elif response.status_code == 200 and request.method in {'GET', 'HEAD'}:
+            # If we're streaming a response and it's small enough we
+            # can buffer it in memory and generate an ETag.
+            streaming = not isinstance(response.app_iter, collections.abc.Sequence)
+            if (streaming and response.content_length is not None and
+                    response.content_length <= BUFFER_MAX):
+                _ = response.body  # noqa
+                streaming = False
+
+            # If we're not streaming the response still we can generate an ETag.
+            if not streaming:
+                response.conditional_response = True
+                response.md5_etag()
+
+        return response
+    return conditional_http_tween
+
+
+def includeme(config):
+    config.add_tween('armonaut.cache.http.conditional_http_tween_factory')